@@ -1,20 +1,21 @@
 package com.super_rabbit.demo.demo_fragments
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.super_rabbit.demo.R
 import com.super_rabbit.demo.wheel_picker_adapters.WPAMPMPickerAdapter
 import com.super_rabbit.demo.wheel_picker_adapters.WPDayPickerAdapter
+import com.super_rabbit.demo.wheel_picker_adapters.WPHoursPickerAdapter
+import com.super_rabbit.demo.wheel_picker_adapters.WPMinutesPickerAdapter
 import com.super_rabbit.wheel_picker.WheelPicker
-import com.syw.domore.adapter.reminder_time_picker_adapters.WPHoursPickerAdapter
-import com.syw.domore.adapter.reminder_time_picker_adapters.WPMinutesPickerAdapter
 
-class DateAndTimePickerFragment : Fragment() {
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
+class DateAndTimePickerFragment : androidx.fragment.app.Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_date_and_time_pikcer, container, false)
     }
@@ -30,8 +31,8 @@ class DateAndTimePickerFragment : Fragment() {
     companion object {
         @JvmStatic
         fun newInstance() =
-                DateAndTimePickerFragment().apply {
-                    arguments = Bundle().apply {}
-                }
+            DateAndTimePickerFragment().apply {
+                arguments = Bundle().apply {}
+            }
     }
 }
