@@ -1,6 +1,6 @@
 package com.super_rabbit.demo.demo_fragments
+
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -8,9 +8,11 @@ import com.super_rabbit.demo.R
 import com.super_rabbit.demo.wheel_picker_adapters.WPWeekDaysPickerAdapter
 import com.super_rabbit.wheel_picker.WheelPicker
 
-class WeekDayPickerFragment : Fragment() {
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
+class WeekDayPickerFragment : androidx.fragment.app.Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_week_day_picker, container, false)
     }
@@ -23,8 +25,8 @@ class WeekDayPickerFragment : Fragment() {
     companion object {
         @JvmStatic
         fun newInstance() =
-                WeekDayPickerFragment().apply {
-                    arguments = Bundle().apply {}
-                }
+            WeekDayPickerFragment().apply {
+                arguments = Bundle().apply {}
+            }
     }
 }
